requirements.txt
Django==4.2
requests==2.28.2
python-dotenv==1.01